+++ apps/api/gql/generated-schema-ast.ts
@@ -224,6 +224,78 @@ export const schema = gql`
         userWalletSyncBlock: BigInt!
     }
 
+    type GqlLoopsData {
+        """
+        Actual TotalSupply of LoopS.
+        """
+        actualSupply: String!
+
+        """
+        The total APR for LoopS
+        """
+        apr: Float!
+
+        """
+        The amount of stS provided to Aave.
+        """
+        collateralAmount: String!
+
+        """
+        The amount of stS provided to Aave in S.
+        """
+        collateralAmountInEth: String!
+
+        """
+        The total S debt amount of the position
+        """
+        debtAmount: String!
+
+        """
+        The health factor of the Aave position
+        """
+        healthFactor: String!
+
+        """
+        The amount of leverage the current position has.
+        """
+        leverage: Float!
+
+        """
+        Net Asset Value. The amount of collateral minus the amount of debt.
+        """
+        nav: String!
+
+        """
+        The current rate of LoopS against S.
+        """
+        rate: String!
+
+        """
+        The current Sonic points multiplier for LoopS
+        """
+        sonicPointsMultiplier: String!
+
+        """
+        The current amount of stS supplied to the Aave market
+        """
+        stSAaveMarketSupply: String!
+
+        """
+        The current cap on the stS market on Aave
+        """
+        stSAaveMarketSupplyCap: String!
+
+        """
+        The health factor that the Aave position should have
+        """
+        targetHealthFactor: String!
+
+        """
+        Net Asset Value in USD.
+        """
+        tvl: String!
+    }
+
     """
     All info on the nested pool if the token is a BPT. It will only support 1 level of nesting.
     """
@@ -670,7 +742,7 @@ export const schema = gql`
     """
     type GqlPoolAprItem {
         """
-        The APR value in % -> 0.2 = 0.2%
+        The APR value in % -> 0.2 = 20%
         """
         apr: Float!
 
@@ -4224,6 +4296,11 @@ export const schema = gql`
         blocksGetBlocksPerYear: Float! @deprecated
         latestSyncedBlocks: GqlLatestSyncedBlocks!
         lbpPriceChart(chain: GqlChain!, dataPoints: Int, id: String!, interval: Int @deprecated): [LBPPriceChartData!]
+
+        """
+        Get the LoopS data
+        """
+        loopsGetData: GqlLoopsData!
 
         """
         Getting swap, add and remove events with paging