+++ apps/api/gql/generated-schema.ts
@@ -218,6 +218,38 @@ export interface GqlLatestSyncedBlocks {
     userWalletSyncBlock: Scalars['BigInt'];
 }
 
+export interface GqlLoopsData {
+    __typename?: 'GqlLoopsData';
+    /** Actual TotalSupply of LoopS. */
+    actualSupply: Scalars['String'];
+    /** The total APR for LoopS */
+    apr: Scalars['Float'];
+    /** The amount of stS provided to Aave. */
+    collateralAmount: Scalars['String'];
+    /** The amount of stS provided to Aave in S. */
+    collateralAmountInEth: Scalars['String'];
+    /** The total S debt amount of the position */
+    debtAmount: Scalars['String'];
+    /** The health factor of the Aave position */
+    healthFactor: Scalars['String'];
+    /** The amount of leverage the current position has. */
+    leverage: Scalars['Float'];
+    /** Net Asset Value. The amount of collateral minus the amount of debt. */
+    nav: Scalars['String'];
+    /** The current rate of LoopS against S. */
+    rate: Scalars['String'];
+    /** The current Sonic points multiplier for LoopS */
+    sonicPointsMultiplier: Scalars['String'];
+    /** The current amount of stS supplied to the Aave market */
+    stSAaveMarketSupply: Scalars['String'];
+    /** The current cap on the stS market on Aave */
+    stSAaveMarketSupplyCap: Scalars['String'];
+    /** The health factor that the Aave position should have */
+    targetHealthFactor: Scalars['String'];
+    /** Net Asset Value in USD. */
+    tvl: Scalars['String'];
+}
+
 /** All info on the nested pool if the token is a BPT. It will only support 1 level of nesting. */
 export interface GqlNestedPool {
     __typename?: 'GqlNestedPool';
@@ -419,7 +451,7 @@ export interface GqlPoolApr {
 /** All APRs for a pool */
 export interface GqlPoolAprItem {
     __typename?: 'GqlPoolAprItem';
-    /** The APR value in % -> 0.2 = 0.2% */
+    /** The APR value in % -> 0.2 = 20% */
     apr: Scalars['Float'];
     /** The id of the APR item */
     id: Scalars['ID'];
@@ -2985,6 +3017,8 @@ export interface Query {
     blocksGetBlocksPerYear: Scalars['Float'];
     latestSyncedBlocks: GqlLatestSyncedBlocks;
     lbpPriceChart?: Maybe<Array<LbpPriceChartData>>;
+    /** Get the LoopS data */
+    loopsGetData: GqlLoopsData;
     /** Getting swap, add and remove events with paging */
     poolEvents: Array<GqlPoolEvent>;
     /**
@@ -3407,6 +3441,7 @@ export type ResolversTypes = ResolversObject<{
     GqlHookType: GqlHookType;
     GqlLBPTopTrade: ResolverTypeWrapper<GqlLbpTopTrade>;
     GqlLatestSyncedBlocks: ResolverTypeWrapper<GqlLatestSyncedBlocks>;
+    GqlLoopsData: ResolverTypeWrapper<GqlLoopsData>;
     GqlNestedPool: ResolverTypeWrapper<GqlNestedPool>;
     GqlPoolAddRemoveEventV3: ResolverTypeWrapper<GqlPoolAddRemoveEventV3>;
     GqlPoolAggregator: ResolverTypeWrapper<GqlPoolAggregator>;
@@ -3638,6 +3673,7 @@ export type ResolversParentTypes = ResolversObject<{
     GqlHookReviewData: GqlHookReviewData;
     GqlLBPTopTrade: GqlLbpTopTrade;
     GqlLatestSyncedBlocks: GqlLatestSyncedBlocks;
+    GqlLoopsData: GqlLoopsData;
     GqlNestedPool: GqlNestedPool;
     GqlPoolAddRemoveEventV3: GqlPoolAddRemoveEventV3;
     GqlPoolAggregator: GqlPoolAggregator;
@@ -3989,6 +4025,27 @@ export type GqlLatestSyncedBlocksResolvers<
     poolSyncBlock?: Resolver<ResolversTypes['BigInt'], ParentType, ContextType>;
     userStakeSyncBlock?: Resolver<ResolversTypes['BigInt'], ParentType, ContextType>;
     userWalletSyncBlock?: Resolver<ResolversTypes['BigInt'], ParentType, ContextType>;
+    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+}>;
+
+export type GqlLoopsDataResolvers<
+    ContextType = ResolverContext,
+    ParentType extends ResolversParentTypes['GqlLoopsData'] = ResolversParentTypes['GqlLoopsData'],
+> = ResolversObject<{
+    actualSupply?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    apr?: Resolver<ResolversTypes['Float'], ParentType, ContextType>;
+    collateralAmount?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    collateralAmountInEth?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    debtAmount?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    healthFactor?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    leverage?: Resolver<ResolversTypes['Float'], ParentType, ContextType>;
+    nav?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    rate?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    sonicPointsMultiplier?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    stSAaveMarketSupply?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    stSAaveMarketSupplyCap?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    targetHealthFactor?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    tvl?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
     __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 }>;
 
@@ -6225,6 +6282,7 @@ export type QueryResolvers<
         ContextType,
         RequireFields<QueryLbpPriceChartArgs, 'chain' | 'id'>
     >;
+    loopsGetData?: Resolver<ResolversTypes['GqlLoopsData'], ParentType, ContextType>;
     poolEvents?: Resolver<
         Array<ResolversTypes['GqlPoolEvent']>,
         ParentType,
@@ -6477,6 +6535,7 @@ export type Resolvers<ContextType = ResolverContext> = ResolversObject<{
     GqlHookReviewData?: GqlHookReviewDataResolvers<ContextType>;
     GqlLBPTopTrade?: GqlLbpTopTradeResolvers<ContextType>;
     GqlLatestSyncedBlocks?: GqlLatestSyncedBlocksResolvers<ContextType>;
+    GqlLoopsData?: GqlLoopsDataResolvers<ContextType>;
     GqlNestedPool?: GqlNestedPoolResolvers<ContextType>;
     GqlPoolAddRemoveEventV3?: GqlPoolAddRemoveEventV3Resolvers<ContextType>;
     GqlPoolAggregator?: GqlPoolAggregatorResolvers<ContextType>;