+++ config/sonic.ts
@@ -50,6 +50,11 @@ export default <NetworkData>{
         baseAprUrl: 'https://apr.soniclabs.com/current',
         validatorFee: 0.15,
     },
+    loops: {
+        address: '0xc76995054ce51dfbbc954840d699b2f33d2538ee',
+        aavePoolDataProvider: '0x9005a69fe088680827f292e8ae885be4be1beb2f',
+        aavePoolAddressesProvider: '0x5c2e738f6e27bce0f7558051bf90605dd6176900',
+    },
     balancer: {
         v2: {
             vaultAddress: '0xba12222222228d8ba445958a75a0704d566bf2c8',